+++ include/engine/guidance/segment_list.hpp
@@ -9,7 +9,7 @@
 #include "engine/segment_information.hpp"
 #include "util/integer_range.hpp"
 #include "util/coordinate_calculation.hpp"
-#include "util/container.hpp"
+#include "util/for_each_pair.hpp"
 #include "extractor/turn_instructions.hpp"
 
 #include <boost/assert.hpp>
